Allow termination of Scheduler thread pools
