+++ core/jvm/src/main/scala/fs2/Scheduler.scala
@@ -1,13 +1,13 @@
 package fs2
 
-import java.lang.Thread.UncaughtExceptionHandler
-import java.util.concurrent.{ Executors, ScheduledExecutorService, ThreadFactory, TimeUnit }
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.{ Executors, ScheduledExecutorService, TimeUnit }
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
 
-import fs2.internal.NonFatal
+import cats.effect.Sync
+
+import fs2.internal.ThreadFactories
 
 /** Provides the ability to schedule evaluation of thunks in the future. */
 trait Scheduler {
@@ -35,10 +35,21 @@ trait Scheduler {
 }
 
 object Scheduler {
+  /** Returns a singleton stream consisting of a `Scheduler`. The scheduler is shutdown at finalization of the stream. */
+  def apply[F[_]](corePoolSize: Int, daemon: Boolean = true, threadPrefix: String = "fs2-scheduler", exitJvmOnFatalError: Boolean = true)(implicit F: Sync[F]): Stream[F,Scheduler] =
+    Stream.bracket(allocate(corePoolSize, daemon, threadPrefix, exitJvmOnFatalError))(t => Stream.emit(t._1), _._2)
 
-  def fromFixedDaemonPool(corePoolSize: Int, threadName: String = "Scheduler.fromFixedDaemonPool"): Scheduler =
-    fromScheduledExecutorService(Executors.newScheduledThreadPool(corePoolSize, daemonThreadFactory(threadName)))
+  /**
+   * Allocates a scheduler in the specified effect type and returns the scheduler along with a shutdown
+   * task, that when executed, terminates the thread pool used by the scheduler.
+   */
+  def allocate[F[_]](corePoolSize: Int, daemon: Boolean = true, threadPrefix: String = "fs2-scheduler", exitJvmOnFatalError: Boolean = true)(implicit F: Sync[F]): F[(Scheduler,F[Unit])] =
+    F.delay {
+      val executor = Executors.newScheduledThreadPool(corePoolSize, ThreadFactories.threadFactory(threadPrefix, daemon, exitJvmOnFatalError))
+      fromScheduledExecutorService(executor) -> F.delay(executor.shutdown)
+    }
 
+  /** Creates a scheduler from the specified executor service. */
   def fromScheduledExecutorService(service: ScheduledExecutorService): Scheduler = new Scheduler {
     override def scheduleOnce(delay: FiniteDuration)(thunk: => Unit): () => Unit = {
       val f = service.schedule(new Runnable { def run = thunk }, delay.toNanos, TimeUnit.NANOSECONDS)
@@ -50,28 +61,4 @@ object Scheduler {
     }
     override def toString = s"Scheduler($service)"
   }
-
-  /** A `ThreadFactory` which creates daemon threads, using the given name. */
-  private[fs2] def daemonThreadFactory(threadName: String, exitJvmOnFatalError: Boolean = true): ThreadFactory = new ThreadFactory {
-    val defaultThreadFactory = Executors.defaultThreadFactory()
-    val idx = new AtomicInteger(0)
-    def newThread(r: Runnable) = {
-      val t = defaultThreadFactory.newThread(r)
-      t.setDaemon(true)
-      t.setName(s"$threadName-${idx.incrementAndGet()}")
-      t.setUncaughtExceptionHandler(new UncaughtExceptionHandler {
-        def uncaughtException(t: Thread, e: Throwable): Unit = {
-          ExecutionContext.defaultReporter(e)
-          if (exitJvmOnFatalError) {
-            e match {
-              case NonFatal(_) => ()
-              case fatal => System.exit(-1)
-            }
-          }
-        }
-      })
-      t
-    }
-  }
-
 }
