+++ core/shared/src/main/scala/fs2/Stream.scala
@@ -1444,9 +1444,11 @@ object Stream {
      *
      * @example {{{
      * scala> import scala.concurrent.duration._, scala.concurrent.ExecutionContext.Implicits.global, cats.effect.IO
-     * scala> implicit val scheduler: Scheduler = Scheduler.fromFixedDaemonPool(1)
-     * scala> val s = Stream(1, 2, 3) ++ time.sleep_[IO](400.millis) ++ Stream(4, 5) ++ time.sleep_[IO](10.millis) ++ Stream(6)
-     * scala> s.debounce(200.milliseconds).runLog.unsafeRunSync
+     * scala> val s2 = Scheduler[IO](1).flatMap { implicit scheduler =>
+     *      |   val s = Stream(1, 2, 3) ++ time.sleep_[IO](400.millis) ++ Stream(4, 5) ++ time.sleep_[IO](10.millis) ++ Stream(6)
+     *      |   s.debounce(200.milliseconds)
+     *      | }
+     * scala> s2.runLog.unsafeRunSync
      * res0: Vector[Int] = Vector(3, 6)
      * }}}
      */
@@ -1550,9 +1552,11 @@ object Stream {
      *
      * @example {{{
      * scala> import scala.concurrent.duration._, scala.concurrent.ExecutionContext.Implicits.global, cats.effect.IO
-     * scala> implicit val scheduler: Scheduler = Scheduler.fromFixedDaemonPool(1)
-     * scala> val s1 = time.awakeEvery[IO](200.millis).scan(0)((acc, i) => acc + 1)
-     * scala> s1.either(time.sleep_[IO](100.millis) ++ s1).take(10).runLog.unsafeRunSync
+     * scala> val s = Scheduler[IO](1).flatMap { implicit scheduler =>
+     *      |   val s1 = time.awakeEvery[IO](200.millis).scan(0)((acc, i) => acc + 1)
+     *      |   s1.either(time.sleep_[IO](100.millis) ++ s1).take(10)
+     *      | }
+     * scala> s.take(10).runLog.unsafeRunSync
      * res0: Vector[Either[Int,Int]] = Vector(Left(0), Right(0), Left(1), Right(1), Left(2), Right(2), Left(3), Right(3), Left(4), Right(4))
      * }}}
      */
@@ -1776,9 +1780,11 @@ object Stream {
      *
      * @example {{{
      * scala> import scala.concurrent.duration._, scala.concurrent.ExecutionContext.Implicits.global, cats.effect.IO
-     * scala> implicit val scheduler: Scheduler = Scheduler.fromFixedDaemonPool(1)
-     * scala> val s1 = time.awakeEvery[IO](400.millis).scan(0)((acc, i) => acc + 1)
-     * scala> s1.merge(time.sleep_[IO](200.millis) ++ s1).take(6).runLog.unsafeRunSync
+     * scala> val s = Scheduler[IO](1).flatMap { implicit scheduler =>
+     *      |   val s1 = time.awakeEvery[IO](400.millis).scan(0)((acc, i) => acc + 1)
+     *      |   s1.merge(time.sleep_[IO](200.millis) ++ s1)
+     *      | }
+     * scala> s.take(6).runLog.unsafeRunSync
      * res0: Vector[Int] = Vector(0, 0, 1, 1, 2, 2)
      * }}}
      */
